feat(RUN-818): extract interceptors from ipc.py
Move IPCInterceptor, InterceptorRegistry, BudgetInterceptor, and
ObserverInterceptor into dedicated interceptors.py module. Update
__init__.py, harness.py, and test imports accordingly.
